packages/sdk-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devbolt"
version = "0.1.0"
description = "Python SDK for DevBolt feature flags"
readme = "README.md"
authors = [
    { name = "DevBolt", email = "support@modelred.ai" }
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
keywords = ["feature-flags", "feature-toggles", "flags", "devbolt"]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "watchdog>=3.0.0",
    "typing-extensions>=4.0.0; python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-pyyaml>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/ModelRed/devbolt"
Documentation = "https://github.com/ModelRed/devbolt"
Repository = "https://github.com/ModelRed/devbolt"
Issues = "https://github.com/ModelRed/devbolt/issues"

[tool.setuptools.packages.find]
include = ["devbolt*"]

[tool.setuptools.package-data]
devbolt = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=devbolt --cov-report=term-missing --cov-report=html"